requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-cleanup==4.0.0
django-filter==2.2.0
django-import-export==2.1.0
django-sitetree==1.14.0
django-solo==1.1.3
django-uuslug==1.2.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
importlib-metadata==1.6.0
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
python-decouple==3.3
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
